Extract the Electron file dialog into a shared util
chat-view's electronDialog wrapper moves to src/utils/file-dialog.ts and
gains a showSaveDialog method so other features can reuse it.
