+++ eslint.config.mjs
@@ -18,7 +18,7 @@ export default tseslint.config(
         },
     },
     {
-        files: ["src/views/chat-view.ts", "src/views/setup-wizard.ts"],
+        files: ["src/views/chat-view.ts", "src/views/setup-wizard.ts", "src/utils/file-dialog.ts"],
         rules: {
             "@typescript-eslint/no-require-imports": "off",
         },
